端末情報追加
