+++ EngineerResume/Presentation/Screen/Profile/Detail/DataSource/ProfileDetailDataSource.swift
@@ -1,0 +1,9 @@
+enum ProfileDetailSection: CaseIterable {
+    case top
+    case main
+}
+
+enum ProfileDetailItem: Hashable {
+    case top(ProfileModelObject?)
+    case main(ProfileModelObject?)
+}
